requirements.txt
gradio
openai
wandb
azure-storage-blob
azure-identity
debugpy
python-dotenv